requirements.txt
datasets==3.1.0
imageio==2.36.1
imgaug==0.4.0
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
torch==2.4.1
torchmetrics==1.6.0
tqdm==4.67.0
transformers==4.46.2
pyarrow